Move risk_score and risk_severity from tags into higher level Detection_Abstract object. Make sure risk_severity is consistent and included only in json content that requires it. Updated serialization logic for detections.json
